add syslog configuration options
Notice that if the options for syslog are considered invalid or an
error comes up before the configuration is parsed, the default facility
and tag will be used. Respectively `mail` and `policyd-geoip`.

Also *not all* facility name from `syslog.h` will be accepted.
Only `mail,auth,authpriv,daemon,user,local0-7` will be accepted.