update domain guessing game
I think the original idea do check for whois records of the domain
was a bit optimistic in what the whois records contain.
To simply check for the last two parts of the client name is
obviously incorrect in case of, for example, domain with ".co.uk".
Also, it should be noted that, apparently the UK domains have their
own Whois records without country.
So, I have implemented two different way to try to guess the
country from the client name.

The first is to reiterate over the parts of the domain name from
the end and see if any of them has a whois record.
The second is to check the top-level domain and see if it's the
same of a blacklisted country (not all the top-level domains are
country iso codes, of course, but if they are...).