update codebase
The original program was my first or second program in _Golang_.
It was working fine for my needs, but going back to the code for
updates I found parts that I didn't like.
This new version is more verbose, but, I guess, this is more in
line with my current style.

I tried not to make it over-complicated and keep the somewhat,
simplicity of the original.

I have introduced some changes from the previous version that are
quite relevant though.

- Restricted the use of the Whois query to an existing whois
client program or nothing.
- Added guessing the country from the top-level domain of the
client name or, optionally, if set in configuration, the sender address.
- I have slightly changed the return actions (not happy happy, so
it may change again)
- Added a configuration element for the reject message.