remove sender address config flag
It's, matter of fact, unnecessary. Checking or not the sender
address can be controlled by placing the policyd line at the
appropriate place in the Postfix configuration.

If, for example, the check is part of the
_smtpd_helo_restrictions_ Postfix will not pass the sender
(because it does not have it yet).
Placing it in the _smtpd_sender_restrictions_ will, on the other,
hand result in checks for the sender domains.